avoid split package warning
during quarkus build and tests
