+++ ruleops-basic/src/test/java/org/drools/ruleops/basic/MutinyIT.java
@@ -1,8 +1,10 @@
-package org.drools.ruleops;
+package org.drools.ruleops.basic;
 
 import javax.inject.Inject;
 
 import io.quarkus.test.junit.QuarkusTest;
+
+import org.drools.ruleops.LevelTrigger;
 import org.junit.jupiter.api.Test;
 
 // not a: @QuarkusIntegrationTest, this is not to test the result of the build, this is IT with minikube/kind
